Add files via upload
Added the rest of the /proc/meminfo into into the ram.c

I did this as a quick fix to make that RAM usage on my bar more accurate. I notices that the RAM in slstatus would constantly grow over time after a lot of use. With wine, or different browsers. Growing much greater than the values found in a htop, top, and /usr/bin/free. After doing some research and calculations I realized that the Sreclaimable memory was causing the the large difference in values.

I modified the calculation. The ram.c files goes in the components directory of your slstatus.

This is essentially a rough draft , but it works VERY well, and gives me the same values that you would get from free -m in the terminal.